backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
